Restructure file hierarchy (...) - Move api utility methods to utils directory
