+++ src/utils/api/websocket-utilities.ts
@@ -1,13 +1,13 @@
-import { websocketURL } from "../config"
+import { websocketURL } from "../../config"
 import { v4 as uuid } from "uuid"
-import { newBroadcast } from "../interfaces/interfaces";
+import { multiplayerBroadcast } from "../../interfaces/interfaces";
 
 export const multiplayer = {
     // We use a unique ID per client so that a client won't respond to messages originating from itself (eg, updating the UI).
     // This should eliminate flickering, particularly when interacting with the UI with higher latency to the server.
     id: uuid(),
     client: new WebSocket(websocketURL!),
-    commandBuffer: new Set<newBroadcast>(),
+    commandBuffer: new Set<multiplayerBroadcast>(),
 
     reconnect() {
         this.client = new WebSocket(websocketURL!)
